Use decorator instead of overwritting helper
